+++ packages/forestadmin-client/src/index.ts
@@ -96,3 +96,4 @@ export { default as SchemaService, SchemaServiceOptions } from './schema';
 export { default as ActivityLogsService, ActivityLogsOptions } from './activity-logs';
 
 export * from './auth/errors';
+export * from './utils/errors';