+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @forestadmin/agent [1.78.7](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.78.6...@forestadmin/agent@1.78.7) (2026-04-24)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/forestadmin-client:** upgraded to 1.39.4
+* **@forestadmin/mcp-server:** upgraded to 1.11.6
+
+## @forestadmin/agent [1.78.6](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.78.5...@forestadmin/agent@1.78.6) (2026-04-22)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/mcp-server:** upgraded to 1.11.5
+
 ## @forestadmin/agent [1.78.5](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.78.4...@forestadmin/agent@1.78.5) (2026-04-22)